go/worker/storage: Refactor metrics
